% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{add_predictions}
\alias{add_predictions}
\title{Add predictions to a meta dataframe}
\usage{
add_predictions(models, features, meta = NULL)
}
\arguments{
\item{models}{List of models from, e.g., train_models}

\item{features}{A dataframe of features that correspond to the
features the models were trained on (but possibly for different
data).}

\item{meta}{Dataframe to add predictions to.  If NULL, just return
a dataframe of the predictions.}
}
\description{
Given a set of models (as a list) add predictions from model to a
passed dataframe
}
