% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hunspell.R
\name{apply_hunspell}
\alias{apply_hunspell}
\title{Force spell-correct}
\usage{
apply_hunspell(
  text,
  additional_words = NULL,
  skip_prefix = NULL,
  threshold = 1,
  to_lower = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{text}{The text to spell check (list of character vectors)}

\item{additional_words}{List of words that should be considered as
spelled correctly.}

\item{skip_prefix}{List of prefixes that if a word has we should
skip over}

\item{threshold}{Don't try to correct anything that only occurs
threshold or less times in corpus.}

\item{to_lower}{If TRUE will keep everything lowercase, regardless
of spelling suggestions.}
}
\value{
Vector of text, spell-corrected we hope.
}
\description{
Try to spell correct unrecognized words as best as able using the
hunspell package.  This will take the first suggestion of hunspell
and just replace the text with the suggestion, word by word.
}
\details{
All words of 2 or 1 character are skipped.
}
