% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_utility_functions.R
\name{ccs_list_table}
\alias{ccs_list_table}
\title{Given a list of CCS models, produce a table of the models}
\usage{
ccs_list_table(model_list)
}
\arguments{
\item{model_list}{A list of CCS models, or the result of a
ccs_list_table() call.}
}
\description{
This is a wrapper of the textreg make.list.table() method with some
extra cleanup.
}
