% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_utility_functions.R
\name{ccs_tuned_textreg}
\alias{ccs_tuned_textreg}
\title{Find threshold C and run textreg with that tuning parameter.}
\usage{
ccs_tuned_textreg(corpus, Z, R = 20, beta = 0.8, ...)
}
\arguments{
\item{R}{The number of randomizations to use for the null
distribution.}

\item{beta}{The quantile of the null distribution to use for
regularization.}

\item{...}{Additional arguments to pass to textreg.  See
textreg::textreg for more information.}
}
\value{
Textreg result object with the passed tuning parameter.
}
\description{
Find threshold C and run textreg with that tuning parameter.
}
