% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Generic text cleaning function}
\usage{
clean_text(x)
}
\arguments{
\item{x}{Character vector or corpus object.}
}
\description{
Quick and easy text cleaning.  Take given copus and
remove punctuation, remove whitespace, convert everything
to lowercase.
}
