% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dimnames}
\alias{dimnames}
\title{Dataset containing word embeddings on 50 dimensions based on GloVe pre-trained embedding model}
\format{
A \link{data.frame} with 36 rows and 2 columns
}
\usage{
dimnames
}
\description{
Dataset containing word embeddings on 50 dimensions based on GloVe pre-trained embedding model
}
\details{
Pretty-fied dimension names for LIWC features
}
\keyword{data}
