% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_impacts.R
\name{estimate_impacts}
\alias{estimate_impacts}
\title{Estimate treatment impacts for hybrid-scored text outcomes}
\usage{
estimate_impacts(
  y.obs,
  yhat,
  Z,
  wts = NULL,
  design = c("crd", "multi", "cluster", "rcbd"),
  siteID = NULL,
  clusterID = NULL,
  data,
  adjust = NULL
)
}
\arguments{
\item{y.obs}{A vector of human-coded scores (with NAs for unscored documents).}

\item{yhat}{A vector of predicted scores estimated via \code{predict_scores}.}

\item{Z}{Indicator for treatment assignment.}

\item{wts}{Sampling weights for which documents were human scored.  Assumed uniform if null.}

\item{design}{Type of design used for random assignment (complete
randomization, multisite randomized, cluster randomized, and blocked and cluster randomized).}

\item{siteID}{Vector of IDs for site, for multi-site randomized experiments.}

\item{clusterID}{Vector of IDs for cluster, for cluster-randomized experiments.}

\item{data}{A \code{data.frame} of subject-level identifiers, demographic
variables, group membership, and/or other pre-treatment covariates.}

\item{adjust}{(optional) character vector or named list of variables in the data matrix to
adjust for when estimating treatment impacts.}
}
\value{
A model object for estimating treatment impact across an array of features.
}
\description{
Given text from a randomized trial with a binary treatment, where a subset of
the documents have been human-scored, this function computes model-assisted
estimates for the average treatment effect with respect to the human-coded
outcome.
}
