% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Generate an array of text features}
\usage{
extract_features(x, p_max = NULL, glove = c(300, 200, 100, 50), ...)
}
\arguments{
\item{x}{A \link{corpus} object or character vector of text documents.}

\item{p_max}{The maximum number of features to compute. Defaults to \code{NULL} (no strict limit).}

\item{glove}{Number of Word2Vec components to compute using GloVe pre-trained embedding model \insertCite{pennington2014glove}{tada}.
Defaults to 300.}

\item{...}{(optional) additional arguments passed to \link{tokens()} for text pre-processing.}
}
\value{
A matrix of available text features, one row per document, one column per feature.
}
\description{
Generates a rich feature representation for documents provided as a character vector or \link{quanteda::corpus()} object
by applying an array of linguistic and syntactic indices, available text analysis dictionaries,
and pre-trained embedding models to all documents.
}
\references{
{
\insertRef{pennington2014glove}{tada}
}
}
