% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_liwc.R
\name{extract_liwc}
\alias{extract_liwc}
\title{Functions for processing and appending output from Linguistic
Inquiry Word Count (LIWC) software}
\usage{
extract_liwc(file, meta = NULL, ID.liwc = 1, ID.meta = NULL, clean = TRUE)
}
\arguments{
\item{file}{character path to LIWC output file}

\item{meta}{optional data frame to attach LIWC output to}

\item{ID.liwc}{ID column (either name or column number) of document
IDs in the liwc file.}

\item{ID.meta}{If meta is specified, character vector of the
document ID column to use for merging, corresponding with IDs in
ID.liwc.}

\item{clean}{should LIWC output be cleaned prior to appending
(e.g., remove linear combinations, repeat variables, etc.)}
}
\description{
Functions for processing and appending output from Linguistic
Inquiry Word Count (LIWC) software
}
