% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taaco.R
\name{extract_taaco}
\alias{extract_taaco}
\title{Manage and merge text features generated using TAACO}
\usage{
extract_taaco(file, data = NULL, idvar = NULL)
}
\arguments{
\item{file}{Filename where TAACO results are stored}

\item{data}{Optional \code{data.frame} with additional document-level variables to include in output.}

\item{idvar}{If \code{data} is specified, character vector with name(s) of variables used for merging.}
}
\value{
Returns a \code{data.frame} of text features.
}
\description{
Tools to support feature extraction using TAACO.
\code{extract_taaco()} reads output and log files produced by TAACO program and returns a \code{data.frame} that can be merged with other feature sets.
}
