% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{glove.50d}
\alias{glove.50d}
\title{Dataset containing word embeddings on 50 dimensions based on GloVe pre-trained embedding model}
\format{
A \link{data.frame} with 400,000 terms and 50 dimensions of word embeddings
}
\usage{
glove.50d
}
\description{
Dataset containing word embeddings on 50 dimensions based on GloVe pre-trained embedding model
}
\details{
50-dimensional word embedding vectors for 400,000 terms and phrases based on GloVe pre-trained embedding model
}
\keyword{data}
