% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfx.R
\name{impacts_on_features}
\alias{impacts_on_features}
\title{For each column of x, conduct an analysis of impact of MORE
intervention on feature represented by that column.}
\usage{
impacts_on_features(
  data,
  meta = NULL,
  ignore = NULL,
  analysis_function = simple_RCT_analysis,
  mcp = "fdr",
  planned_features = NULL,
  standardize = TRUE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{Dataframe of the meta information and features.  Will
calculate impacts on all columns of data that are not specified
by ignore.}

\item{meta}{Extra information that are not text features to be
analyzed (e.g. treatment assignment indicator) that should be
given to each analysis for each feature (this is a dataframe)}

\item{ignore}{Character vector of column names to ignore in the
analysis.  Or list of column indices.}

\item{analysis_function}{Function to apply to each column of data.
Should return a one-row tibble with columns such as Estimate, SE,
t, p, CI_l, CI_h.  Function is of form f( feature, data, ... )
with feature being a vector of values, with length equal to
number of rows of data.  Function results MUST return a column
'p.value' of the p.value of the test.}

\item{mcp}{character string specifying the correction method to be
applied to adjust for multiple comparisons. Defaults to fdr
adjustments. See \link{p.adjust} for available adjustment
methods.  NULL means no adjustment.}

\item{planned_features}{Character vector of features that are
planned for assessment.  They will never be screened out,
regardless of their p-value.  Multiple comparison is done within
group of planned and unplanned features.}

\item{standardize}{Logical.  If TRUE, standardize the effect sizes
in new columns of output with suffix "_std".  If "pool", pool the
standard deviations of the two treatment groups to get the
standardize metric.}

\item{alpha}{Significance level for hypothesis tests.}

\item{...}{Additional arguments to pass to analysis_function.}
}
\description{
Adjust all tests with FDR at end.  For an example function for
analysis_function see simple_RCT_analysis.
}
\seealso{
simple_RCT_analysis
}
