% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liwc.R
\name{liwc}
\alias{liwc}
\title{Functions for processing and appending output from Linguistic Inquiry Word Count (LIWC) software}
\usage{
liwc(file, x = NULL, by = NULL, clean = T)
}
\arguments{
\item{file}{character path to LIWC output file}

\item{x}{optional data frame to append LIWC output}

\item{by}{if x is specified, character vector of the columns used for merging}

\item{clean}{should LIWC output be cleaned prior to appending (e.g., remove linear combinations, repeat variables, etc.)}
}
\description{
Functions for processing and appending output from Linguistic Inquiry Word Count (LIWC) software
}
