% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mini_glove}
\alias{mini_glove}
\title{Mini glove dataset}
\format{
A \link{matrix} with 1000 rows and 50 columns
}
\usage{
mini_glove
}
\description{
glove embeddings (50 dimensional) for 1000 common words beyond
those words listed in several stopword lists provided by quanteda.
}
\details{
Original glove embeddings downloaded from Stanford CITE.

1000 by 50 matrix, each row is an embedding.  Rownames are
  words.
}
\keyword{data}
