% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_distances.R
\name{pairwise_distances}
\alias{pairwise_distances}
\title{Calculate all pairwise distances between documents in two corpus.}
\usage{
pairwise_distances(
  corpus1,
  corpus0,
  method = c("cosine", "w2v"),
  wide = TRUE,
  model = NULL,
  use_names = TRUE,
  names_prefix = "doc_"
)
}
\arguments{
\item{wide}{If TRUE return distances with each column corresponding
to a document in corpus0, and each row a document in corpus1.  If
FALSE return dataframe of indices for each pair, with the third
column being distance.}

\item{model}{When using w2v, a user-specified model object pointing
to a pre-trained embedding model, represented as a data frame
where the first column is the word/token and the following
columns are numeric vectors.}
}
\value{
Dataframe of either length(corpus1) rows or
  length(corpus1)*length(corpus0) rows, depending on wide.
}
\description{
Calculate all pairwise distances between documents in two corpus.
}
