% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ccs.R
\name{plot.ccs}
\alias{plot.ccs}
\title{Plot the results from a CCS run}
\usage{
\method{plot}{ccs}(out, xadj = c(-0.025, 0.025), ...)
}
\arguments{
\item{out}{a \link{textreg.result()} object}

\item{xadj}{adjustments to the lower and upper limits on the x-axis of the plot}

\item{...}{additional arguments passed to plot}
}
\description{
This function provides a visualization of the set of words and phrases
found to differ systematically between treatment and control groups
}
