% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.textfx.R
\name{plot.textfx}
\alias{plot.textfx}
\title{Plot the results from an impact analysis with text outcomes}
\usage{
\method{plot}{textfx}(out, alpha = 0.05, cols = F, group = NULL, ...)
}
\arguments{
\item{alpha}{the threshold for determining statistical significance}

\item{cols}{should effects be colored by direction (red for negative impacts, blue for positive impacts)}

\item{group}{(optional) should effects be grouped by category (e.g., higher-level summary measures, linguistic features, etc.)}

\item{...}{additional arguments passed to plot}

\item{x}{a model object output from \code{estimate_impacts()}}
}
\description{
This function provides a visualization of the set of textual features
found to differ systematically between treatment and control groups.
}
