% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impacts.R
\name{plot_impacts}
\alias{plot_impacts}
\title{Plot the results from an impact analysis with text outcomes}
\usage{
plot_impacts(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{a model object output from \code{estimate_impacts()}}

\item{alpha}{the threshold for determining statistical significance}

\item{...}{additional arguments passed to plotting method.}
}
\description{
This function provides a visualization of the set of textual features
found to differ systematically between treatment and control groups.
}
