% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-documentation.R
\docType{package}
\name{rcttext}
\alias{rcttext}
\alias{rcttext-package}
\title{The `rcttext` package}
\description{
The `rcttext` package provides a set of functions for working
  with text data in R, when that text is an outcome in a randomized
  controlled trial. The package includes functions for cleaning and
  preprocessing text data, as well as for applying text mining
  techniques such as sentiment analysis and topic modeling.
}
\author{
\strong{Maintainer}: Reagan Mozer \email{rmozer@g.harvard.edu} [copyright holder]

Authors:
\itemize{
  \item Luke Miratrix \email{lmiratrix@g.harvard.edu} [copyright holder]
  \item Shireen Al-Adeimi \email{aladeimi@msu.edu} [data contributor]
}

}
