% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_spelling.R
\name{repair_spelling}
\alias{repair_spelling}
\title{Replace all words in dictionary with alternates}
\usage{
repair_spelling(text, dictionary, to_words = NULL)
}
\arguments{
\item{text}{Character vector}

\item{dictionary}{Dataframe with two columns of text, or a list of
words.}

\item{to_words}{If dictionary is a list of words, this is list of
corresponding words.}

\item{Character}{vector, revised version of text.}
}
\description{
Given text as a list of character strings, and a dictionary as a
two-column dataframe with the first column being misspelled words
and the second being corect spelling, swap all misspelled words
with the correct spellings.
}
