% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccs.R
\name{results.tab}
\alias{results.tab}
\title{Make results table for grid CCS run}
\usage{
results.tab(result, corp, Z)
}
\arguments{
\item{result}{a \link{textreg.result()} object}

\item{corp}{a corpus or character vector to calculate term frequencies across}

\item{Z}{an indicator for treatment assignment}

\item{clusterID}{optional vector of cluster ID's}

\item{...}{additional arguments passed to \link{textreg()}.}
}
\value{
a \link{textreg.result()} object.
}
\description{
Make results table for grid CCS run
}
