% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccs.R
\name{run_ccs}
\alias{run_ccs}
\alias{cluster.threshold.C}
\title{Perform Concise Comparative Summarization}
\usage{
run_ccs(x, z, ...)

cluster.threshold.C(
  x,
  z,
  design = c("crd", "multi", "cluster", "rcbd"),
  clusterID,
  siteID = NULL,
  R,
  ...
)
}
\arguments{
\item{x}{a corpus, character vector of text documents, or set of text features.}

\item{z}{an indicator for treatment assignment}

\item{...}{additional arguments passed to \link{textreg()}.}

\item{design}{Type of design used for random assignment (complete
randomization, multisite randomized, cluster randomized, and blocked and cluster randomized).}

\item{clusterID}{vector of cluster ID's}

\item{siteID}{vector of block ID's}

\item{R}{Number of times to scramble treatment assignment labels}

\item{C}{The regularization term. 0 is no regularization.}
}
\value{
a \link{textreg.result()} object.

List of numbers.  First is the threshold C for the passed labeling.
  Remainder are the reference distribution based on the permutations.
}
\description{
Wrapper for \link{textreg::textreg()}.

Determine the penalty C that will zero out the textreg model for a series of
randomly permuted labelings with random assignment dictated by a blocked and
cluster-randomized experiment.
}
\details{
Method repeatedly generates +1/-1 vectors within the given blocking structure
with blocks of +1/-1 within the clustering vector, and then finds a threshold
C for each permutation.
}
