% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_power_study.R
\name{run_power_study}
\alias{run_power_study}
\title{Simulate power under different text scoring configurations.}
\usage{
run_power_study(
  n.coded = NULL,
  n.auto = NULL,
  d = NULL,
  power = 0.8,
  sig.level = 0.05
)
}
\arguments{
\item{n.coded}{Number of hand-coded documents}

\item{n.auto}{Number of additional documents to be scored by ML model.}

\item{d}{Standardized effect size}

\item{power}{Power of test (1 minus Type II error probability)}

\item{sig.level}{Significance level (Type I error probability)}

\item{n.sim}{Number of simulations used for power calculations}
}
\value{
Object of class \code{power.textML}
}
\description{
This function...
}
\keyword{internal}
