% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfx.R
\name{simple_RCT_analysis}
\alias{simple_RCT_analysis}
\title{Default analysis_function for impacts_on_features}
\usage{
simple_RCT_analysis(feature, formula, data, ...)
}
\arguments{
\item{feature}{Vector of values for the feature of interest.}

\item{formula}{Formula to use in the analysis of form ~ Z + X1,
where the first argument has to be the treatment indicator.}

\item{data}{Dataframe with the feature and other variables.}

\item{...}{Additional arguments to pass to lm_robust.}
}
\description{
This is the default analysis function to use in impacts_on_features.
}
\seealso{
impacts_on_features
}
