% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{student_essays}
\alias{student_essays}
\title{MORE study science essays}
\format{
A \link{data.frame} with 40 observations of 4 variables:
\describe{
\item{studentid}{Student identification number}
\item{text}{Student generated essay text}
\item{treatment}{MORE intervention treatment indicator (0=Control, 1=Treatment)}
\item{score}{Human-coded writing quality score}
}
}
\usage{
student_essays
}
\description{
Dataset consisting of a sample of 40 student essays and corresponding human-coded quality scores (sampled from MORE study pilot data)
}
\keyword{data}
