% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tada.R
\name{tada}
\alias{tada}
\title{Generate an array of text features}
\usage{
tada(
  x,
  lex = TRUE,
  sent = TRUE,
  ld = "all",
  read = c("ARI", "Coleman", "DRP", "ELF", "Flesch", "Flesch.Kincaid",
    "meanWordSyllables"),
  terms = NULL,
  preProc = list(uniqueCut = 1, freqCut = 99/1, cor = 0.95, remove.lc = TRUE)
)
}
\arguments{
\item{x}{A \link{corpus} object or character vector of text documents.}

\item{lex}{Logical, indicating whether to compute lexical indices including measures of lexical diversity, readability,
and entropy}

\item{sent}{Logical, indicating whether to compute sentiment analysis features from available dictionaries}

\item{ld}{character vector defining lexical diversity measures to compute; see \link{quanteda.textstats::textstat_lexdiv()}}

\item{read}{character vector defining readability measures to compute; see \link{quanteda.textstats::textstat_readability()}}

\item{terms}{character vector of terms to evaluate as standalone features based on document-level frequency (case-insensitive)}

\item{preProc}{Named list of arguments passed to \code{caret::preProcess()} for applying pre-processing transformations across the set of text features (e.g., removing collinear features)}

\item{...}{(optional) additional arguments passed to \link{quanteda::tokens()} for text pre-processing.}
}
\value{
A matrix of available text features, one row per document, one column per feature.
}
\description{
Generates a rich feature representation for documents provided as a character vector or \link{quanteda::corpus()} object
by applying an array of linguistic and syntactic indices, available text analysis dictionaries,
and pre-trained embedding models to all documents.
}
\references{
{
\insertRef{pennington2014glove}{tada}
}
}
