% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfx.R
\name{textfx_terms}
\alias{textfx_terms}
\title{Estimates the average treatment effect on the frequency and prevalence of a list of specified terms and phrases}
\usage{
textfx_terms(x, Z, terms, ...)
}
\arguments{
\item{x}{A character vector of text documents or a feature matrix}

\item{Z}{Indicator for treatment assignment.}

\item{terms}{Terms and phrases to evaluate}

\item{...}{optional parameters passed to \link{quanteda::tokens()}}
}
\value{
A vector showing the frequency and prevalence of the specified terms within each treatment group and results of the hypothesis test comparing prevalence across groups.
}
\description{
Estimates the average treatment effect on the frequency and prevalence of a list of specified terms and phrases
}
