% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textML.R
\name{textML}
\alias{textML}
\title{Model-assisted impact analysis through hybrid human/machine text scoring}
\usage{
textML(
  x,
  y,
  z = NULL,
  wts = NULL,
  design = c("crd", "multi", "cluster", "rcbd"),
  siteID = NULL,
  clusterID = NULL,
  max.features = NULL,
  ...
)
}
\arguments{
\item{x}{a corpus or character vector of text documents.}

\item{y}{a vector of human-coded scores. Set elements to `NA` for documents
not previously scored.}

\item{z}{optional indicator for treatment assignment. If specified, separate
ensembles will be trained for each treatment group;}

\item{wts}{Sampling weights for which documents were human scored.  Assumed uniform if null.}

\item{design}{Type of design used for random assignment (complete
randomization, multisite randomized, cluster randomized, and blocked and cluster randomized).}

\item{siteID}{Vector of IDs for site, for multi-site randomized experiments.}

\item{clusterID}{Vector of IDs for cluster, for cluster-randomized experiments.}

\item{max.features}{maximum number of text features to use for model
training. Defaults to `NULL` (no strict limit)}

\item{...}{additional arguments passed to \link{train}.}
}
\value{
a \code{textML} model object
}
\description{
A wrapper function for the multiple steps of generating features, training a
scoring model on the human-coded data, predicting scores, and comparing human
v. machine estimates.
}
\details{
This function takes in a corpus of text documents (or a set of computed text
features) along with a sample of human-coded outcome values, and trains an
ensemble of machine learning models to predict the outcome as a function of
the machine measures of text.
}
