% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textsamp.R
\name{textsamp}
\alias{textsamp}
\alias{textsamp_strata}
\alias{textsamp_cluster}
\title{Select a random sample of documents}
\usage{
textsamp(
  x,
  size = length(x),
  prob = NULL,
  wt.fn = NULL,
  scheme = NULL,
  method = c("srswr", "srswor", "systematic", "poisson")
)

textsamp_strata(x, by = NULL, ...)

textsamp_cluster(x, by = NULL, ...)
}
\arguments{
\item{x}{A \link{corpus} object or character vector of text documents.}

\item{size}{a non-negative integer giving the number of documents to sample.}

\item{prob}{a vector of probability weights for each document.}

\item{wt.fn}{a function for generating probability weights; ignored when \code{prob} is used. See Details.}

\item{scheme}{optional sampling scheme to implement}

\item{method}{the following methods are implemented: simple random sampling without replacement (`srswor`),
simple random sampling with replacement (`srswr`), Poisson sampling (`poisson`), systematic sampling (`systematic`);
if \code{method} is missing, the default method is \code{srswor}.}

\item{by}{a \code{data.frame} with document-level grouping variable(s) or character vector with names of variables in `docvars(x)`}

\item{...}{additional arguments passed on to `textsamp`. Cannot include `scheme`.}
}
\value{
Returns a \code{data.frame} containing identifiers for the selected documents.
}
\description{
Functions to select random samples of documents using different sampling schemes and/or
along different design criteria.
}
