% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{toy_reads}
\alias{toy_reads}
\title{Dataset with 20 essays from READS pilot data}
\format{
A \link{data.frame} with 5 columns and 20 rows
}
\usage{
toy_reads
}
\description{
Used for testing and illustation of rcttext functions.
}
\details{
5 column data.frame, ID is the id of subject, Q1, Q2, more are meta information on scoring, and text contains character string of the text of the essay.
}
\keyword{data}
