% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ensemble.R
\name{train_ensemble}
\alias{train_ensemble}
\title{Train an ensemble learner for semi-supervised text scoring}
\usage{
train_ensemble(
  X,
  Z,
  Yobs,
  coded,
  n.tune = 3,
  bounds = NULL,
  preProc = "zv",
  model = "rf",
  seeds = NA,
  ...
)
}
\arguments{
\item{n.tune}{an integer denoting the amount of granularity in the tuning parameter grid.
By default, this argument is the number of levels for each tuning parameters that should be generated by \link{train}.}

\item{bounds}{a vector (y1, y2) specifying the lower and upper limits for prediction}

\item{...}{additional arguments passed to \link{trainControl}.}

\item{x}{a \code{data.frame} or matrix of numeric text features.}

\item{y}{a vector of human-coded scores for the outcome of interest.}

\item{z}{optional indicator for treatment assignment. If specified, separate ensembles will
be trained for each treatment group;}

\item{cvf}{number of folds for cross validation}

\item{return.all}{should all component models be returned? If `FALSE`, returns only the fitted ensemble(s).}
}
\value{
a fitted model object
}
\description{
This function takes in a corpus of text documents or a set of computed text features, along with a sample
of human-coded outcome values and trains an ensemble of machine learning models to predict the outcome as a
function of machine measures of text.
}
