% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_models}
\alias{train_models}
\title{Train machine learners on a single (unlabeled) dataset.}
\usage{
train_models(
  x,
  y,
  n.tune = 3,
  preProc = NULL,
  bounds = NULL,
  methods = "small",
  include_BART = TRUE,
  include_ensemble = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A matrix of features}

\item{y}{A vector of outcomes}

\item{n.tune}{The number of tuning parameters to search over}

\item{preProc}{A pre-processing list}

\item{bounds}{The bounds of the outcome variable}

\item{methods}{A character vector of methods to use.  Special
values of "small" will use a small, hand-picked list of defaults.
"full" will use a larger list of defaults.}

\item{include_BART}{Should we include BART in the ensemble?}

\item{include_ensemble}{Should we include the ensemble in the list
of models returned?  If TRUE, will return vanilla list, not a
caretList, of models.}

\item{verbose}{Should we print out progress?}
}
\value{
A list of caret 'train' objects (or ensamble equivilents)
  that can be used to predict on new data.
}
\description{
Train a bunch of machine learners, along with an ensemble made of
those learners, and return the list of fit models.
}
