% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hunspell.R
\name{apply_hunspell}
\alias{apply_hunspell}
\title{Force spell-correct}
\usage{
apply_hunspell(
  text,
  additional_words = NULL,
  skip_prefix = NULL,
  threshold = 1,
  to_lower = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{text}{The text to spell check (list of character vectors)}

\item{additional_words}{List of words that should be considered as
spelled correctly.}

\item{skip_prefix}{List of prefixes that if a word has we should
skip over}

\item{threshold}{Don't try to correct anything that only occurs
threshold or less times in corpus.}

\item{to_lower}{If TRUE will keep everything lowercase, regardless
of spelling suggestions.}
}
\value{
Vector of text, spell-corrected we hope.
}
\description{
Try to spell correct unrecognized words as best as able using the
hunspell package.  This will take the first suggestion of hunspell
and just replace the text with the suggestion, word by word.
}
\details{
All words of 2 or 1 character are skipped.
}
\examples{

## Example 1: Single string correction

# Text to be checked for spelling errors
txt = "This function seplaces the textt withh the suggestion"

# Check for spelling errors and apply replacements
txt_rep = apply_hunspell( txt, verbose=TRUE, threshold = 0 )
txt_rep

## Example 2: Multiple texts correction

# Load example dataframe with multiple texts
data("toy_reads")

# Extract texts from the dataframe
txts = toy_reads$text
length(txts)

# Check for spelling errors and apply replacements
txts_rep = apply_hunspell( txts, verbose=TRUE, threshold = 0 )
which( txts_rep != txts )

# Display original and corrected texts
txts[7]
txts_rep[7]

txts[20]
txts_rep[20]

# Check for spelling errors and apply replacements with additional accepted words
txts_rep2 = apply_hunspell( txts,
                            verbose=TRUE,
                            additional_words = c( "dont", "rainforest" ),
                            threshold = 0 )
                            which( txts_rep2 != txts )

# Display original and corrected texts with additional accepted words
txts[7]
txts_rep2[7]

txts[20]
txts_rep2[20]

}
