% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Generic text cleaning function}
\usage{
clean_text(x, split_hyphens = TRUE)
}
\arguments{
\item{x}{Character vector or corpus object.}

\item{split_hyphens}{A logical indicating whether hypenated words should
be treated as two tokens (split at the hypen).}
}
\description{
Quick and easy text cleaning.  Take given copus and
remove punctuation, remove whitespace, convert everything
to lowercase. This function is used for pre-processing text
within the generate_features function. The main use for
clean_text is to check which elements of your text
are converted to empty strings. Empty strings may result in clean_features
dropping desired columns.
}
