% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dimnames}
\alias{dimnames}
\title{Dimension names of LIWC/TAACO that are more accessible}
\format{
A \link{data.frame} with 36 rows and 2 columns
}
\usage{
dimnames
}
\description{
Dimension names of LIWC/TAACO that are more accessible
}
\details{
Pretty-fied dimension names for LIWC/TAACO features
}
\keyword{data}
