% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_w2v.R
\name{extract_w2v}
\alias{extract_w2v}
\title{Compute document-level feature vectors from a pre-trained embedding model.}
\usage{
extract_w2v(x, glove = TRUE, dim = c(50, 100, 200, 300), model = NULL)
}
\arguments{
\item{x}{A \link{corpus()} object or character vector of text documents.}

\item{glove}{Logical.  True means use pre-trained GloVe embedding model.}

\item{dim}{Dimension(s) of the pre-trained GloVe embedding model. Ignored if
\code{glove=FALSE}.}

\item{model}{User-specified model object pointing to a custom pre-trained
Word2Vec model.}
}
\value{
A list of data frames containing the Word2Vec projections of the
  corpus
}
\description{
This function generates a vector embedding for each word in a string using a
set of pre-trained word vectors such as GloVe
\insertCite{pennington2014glove}{tada} and returns the mean vector projection
across all words in a document.
}
\references{
{ \insertRef{mikolov2013efficient}{tada}
\insertRef{pennington2014glove}{tada} }
}
\keyword{internal}
