% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_distances.R
\name{generate_distance_features}
\alias{generate_distance_features}
\title{Generate distances to a set of reference documents}
\usage{
generate_distance_features(
  x,
  meta = NULL,
  reference_docs,
  verbose = FALSE,
  method = c("cosine", "w2v"),
  model = NULL,
  use_names = TRUE,
  names_prefix = "doc_"
)
}
\arguments{
\item{x}{A \link{corpus} object or character vector of text
documents.}

\item{meta}{Dataframe corresponding to the corpus x.  If passed,
and non-NULL, all features will be added to this dataframe.}

\item{reference_docs}{A list of reference documents to compare
corpus against.  Will end up with one distance feature per
reference doc.}

\item{model}{When using w2v, a user-specified model object pointing
to a pre-trained embedding model, represented as a data frame
where the first column is the word/token and the following
columns are numeric vectors.}
}
\description{
This is a helper function for the pairwise_distances method.
}
\seealso{
\code{\link{pairwise_distances}}
}
