% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccs.R
\name{plot_ccs}
\alias{plot_ccs}
\title{Plot the results from a CCS run}
\usage{
plot_ccs(
  out,
  xlim = NULL,
  xadj = c(-0.025, 0.025),
  xlab = "Difference in Usage Rates",
  sub = "(Treatment - Control)",
  ...
)
}
\arguments{
\item{out}{a \link{textreg.result()} object}

\item{xlim}{limits for x-axis}

\item{xadj}{adjustments to the lower and upper limits on the x-axis of the plot}

\item{...}{additional arguments passed to plot}
}
\description{
This function provides a visualization of the set of words and phrases
found to differ systematically between treatment and control groups
}
