% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_scores.R
\name{predict_scores}
\alias{predict_scores}
\title{Extract predictions from a fitted text scoring model.}
\usage{
predict_scores(fit, newdata, na.action = na.omit, ...)
}
\arguments{
\item{fit}{a model or list of models to use for prediction}

\item{newdata}{an optional data frame or matrix of predictors}

\item{na.action}{the method for handling missing data}

\item{...}{additional arguments to pass to \code{predict.train}}
}
\value{
A vector of predictions
}
\description{
This function computes the predicted scores for a collection of documents based on the
results of a trained ensemble learner.
}
